Killing the plugin after wait returns
